Rename the useAutoCompletionState into the useAutoCompletionStateManager
