+++ src/app/App.tsx
@@ -1,11 +1,11 @@
-import { lazy, useEffect } from 'react';
+import { useEffect } from 'react';
 
 import { useLDClient } from 'launchdarkly-react-client-sdk';
 
 import Providers from './providers';
 import i18nManager from './system/locale/i18n';
 
-const ApplicationRouter = lazy(() => import('~/pages'));
+import ApplicationRouter from '~/pages';
 import { Mixpanel } from '~/shared/utils';
 import { FeatureFlags } from '~/shared/utils/featureFlags';
