+++ src/features/AutoCompletion/model/hooks/useAutoCompletionStateManager.ts
