+++ src/features/AutoCompletion/model/index.ts
@@ -1,4 +1,4 @@
 export * from './slice';
 export * from './selectors';
-export * from './hooks/useAutoCompletionState';
+export * from './hooks/useAutoCompletionStateManager';
 export * from './hooks/useAutoCompletion';
